% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{colordew_palette}
\alias{colordew_palette}
\title{Colordew Palette}
\usage{
colordew_palette(name, n, direction = 1, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{name of the desired palette}

\item{n}{number of points to include in the palette}

\item{direction}{either -1 or 1, where if -1 the palette will be reversed}

\item{type}{either "continuous" or "discrete," where "continuous" will automatically
interpolate between colors}
}
\value{
a vector of colors of length n
}
\description{
Returns a named color palette from \code{colordew}
}
\examples{
colordew_palette("winter")
}
\keyword{colors}
