% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_colordew_d}
\alias{scale_color_colordew_d}
\alias{scale_fill_colordew_d}
\alias{scale_color_colordew_c}
\alias{scale_fill_colordew_c}
\title{Discrete colordew scales to be used for ggplot2 fill}
\usage{
scale_color_colordew_d(name, direction = 1, ...)

scale_fill_colordew_d(name, direction = 1, ...)

scale_color_colordew_c(name, direction = 1, ...)

scale_fill_colordew_c(name, direction = 1, ...)
}
\arguments{
\item{name}{name of the desired palette}

\item{direction}{either -1 or 1, where if -1 the palette will be reversed}

\item{...}{additional arguments to pass to continuous_scale()}
}
\description{
Discrete colordew scales to be used for ggplot2 fill

Discrete colordew scales to be used for ggplot2 color

Continuous colordew scales to be used for ggplot2 color

Continuous colordew scales to be used for ggplot2 fill
}
\examples{
library(ggplot2)
days_in_winter_seeds <-
   data.frame(stage = c("stage1", "stage2", "stage3"),
              days = c(3, 4, 7))
ggplot(days_in_winter_seeds, aes(x = stage, y = days, color = stage)) +
   geom_bar(stat = "identity", fill = "white") +
   scale_color_colordew_d("winter")
library(ggplot2)
days_in_winter_seeds <-
   data.frame(stage = c("stage1", "stage2", "stage3"),
              days = c(3, 4, 7))
ggplot(days_in_winter_seeds, aes(x = stage, y = days, fill = stage)) +
   geom_bar(stat = "identity") +
   scale_fill_colordew_d("winter")
library(ggplot2)
days_in_winter_seeds <-
  data.frame(stage = c(1, 2, 3),
             cumulative_days = c(3, 7, 14))
ggplot(days_in_winter_seeds, aes(x = stage, y = cumulative_days, color = stage)) +
  geom_line() +
  scale_color_colordew_c("winter")
days_in_winter_seeds <-
  data.frame(stage = c(1, 2, 3),
             cumulative_days = c(3, 7, 14))
ggplot(days_in_winter_seeds, aes(x = stage,
                                 y = cumulative_days,
                                 fill = stage)) +
 geom_point(shape = 22) +
 scale_fill_colordew_c("winter")
}
